# Use a direct cache in index reader for label values