A direct cache shows significant speedup in both the benchmarks (below),
as well as real life data. In real life data I am seeing Thanos style
compaction go from 36 minutes to 25 minutes. Thanos does two passes on
the index read making this optimization more beneficial than the
Prometheus benchmarks.

```
name                                                                          old time/op    new time/op    delta
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-8      404ms ± 5%     308ms ± 4%  -23.82%  (p=0.000 n=8+7)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-8     735ms ± 1%     632ms ± 2%  -14.00%  (p=0.000 n=8+8)

name                                                                          old alloc/op   new alloc/op   delta
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-8     86.3MB ± 0%    80.6MB ± 0%   -6.58%  (p=0.000 n=8+8)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-8     125MB ± 0%     120MB ± 0%   -4.39%  (p=0.000 n=8+8)

name                                                                          old allocs/op  new allocs/op  delta
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=101-8      1.33M ± 0%     0.97M ± 0%  -27.03%  (p=0.000 n=8+8)
Compaction/type=normal,blocks=4,series=10000,samplesPerSeriesPerBlock=1001-8     2.29M ± 0%     1.93M ± 0%  -15.70%  (p=0.000 n=8+8)
```

<!--
    Don't forget!
    
    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.
    
    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.
    
    - No tests are needed for internal implementation changes.
    
    - Performance improvements would need a benchmark test to prove it.
    
    - All exposed objects should have a comment.
    
    - All comments should start with a capital letter and end with a full stop.
 -->