@brian-brazil 2020-12-15
What does this show for the querier benchmarks? What is the memory and memory allocation impact?

A change like this should be evaluated primarily by its effect on queries, not compaction. Compaction has atypical access patterns, and we can always add code to make things better there in a way that doesn't risk query performance. These microbencharks are also not a good representation of real-world performance due to how the fake data is generates, so the gains shown here won't translate to reality.

We already recently added a cache that I think should give us all the gains we can get here in terms of compation, what code path is this mean to be helping with?