@csmarchbanks 2020-12-15
Good idea to check the querier benchmarks as well. The QuerierSelect benchmarks also show a reasonable speedup, and a reduction in allocations. Are there other benchmarks you would like me to run?

```
name                                    old time/op    new time/op    delta
QuerierSelect/Block/1of1000000-8           552ms ± 1%     523ms ± 0%   -5.37%  (p=0.008 n=5+5)
QuerierSelect/Block/10of1000000-8          559ms ± 2%     524ms ± 1%   -6.26%  (p=0.008 n=5+5)
QuerierSelect/Block/100of1000000-8         554ms ± 2%     522ms ± 1%   -5.68%  (p=0.008 n=5+5)
QuerierSelect/Block/1000of1000000-8        553ms ± 1%     526ms ± 1%   -4.91%  (p=0.008 n=5+5)
QuerierSelect/Block/10000of1000000-8       556ms ± 0%     523ms ± 1%   -5.93%  (p=0.016 n=4+5)
QuerierSelect/Block/100000of1000000-8      569ms ± 0%     536ms ± 1%   -5.64%  (p=0.008 n=5+5)
QuerierSelect/Block/1000000of1000000-8     723ms ± 2%     687ms ± 2%   -4.98%  (p=0.008 n=5+5)

name                                    old alloc/op   new alloc/op   delta
QuerierSelect/Block/1of1000000-8          99.2MB ± 0%    96.0MB ± 0%   -3.22%  (p=0.008 n=5+5)
QuerierSelect/Block/10of1000000-8         99.2MB ± 0%    96.0MB ± 0%   -3.22%  (p=0.008 n=5+5)
QuerierSelect/Block/100of1000000-8        99.2MB ± 0%    96.0MB ± 0%   -3.22%  (p=0.008 n=5+5)
QuerierSelect/Block/1000of1000000-8       99.3MB ± 0%    96.1MB ± 0%   -3.22%  (p=0.008 n=5+5)
QuerierSelect/Block/10000of1000000-8       100MB ± 0%      97MB ± 0%   -3.18%  (p=0.008 n=5+5)
QuerierSelect/Block/100000of1000000-8      112MB ± 0%     109MB ± 0%   -2.85%  (p=0.000 n=4+5)
QuerierSelect/Block/1000000of1000000-8     227MB ± 0%     224MB ± 0%   -1.41%  (p=0.008 n=5+5)

name                                    old allocs/op  new allocs/op  delta
QuerierSelect/Block/1of1000000-8           3.00M ± 0%     2.00M ± 0%  -33.30%  (p=0.008 n=5+5)
QuerierSelect/Block/10of1000000-8          3.00M ± 0%     2.00M ± 0%  -33.30%  (p=0.008 n=5+5)
QuerierSelect/Block/100of1000000-8         3.00M ± 0%     2.00M ± 0%  -33.30%  (p=0.008 n=5+5)
QuerierSelect/Block/1000of1000000-8        3.00M ± 0%     2.00M ± 0%  -33.28%  (p=0.008 n=5+5)
QuerierSelect/Block/10000of1000000-8       3.02M ± 0%     2.02M ± 0%  -33.08%  (p=0.008 n=5+5)
QuerierSelect/Block/100000of1000000-8      3.20M ± 0%     2.20M ± 0%  -31.22%  (p=0.008 n=5+5)
QuerierSelect/Block/1000000of1000000-8     5.00M ± 0%     4.00M ± 0%  -19.98%  (p=0.008 n=5+5)
```

As far as compaction on real life data goes, as I said in my commit/PR description, when testing on real data I am also seeing a nice improvement in compaction times. The cache that was recently added was only for the write path for the index, this is on the read path. Specifically, the `index.(*Decoder).Series` calls inside of `storage.(*genericMergeSeriesSet).Next` are the ones being optimized. You can compare the two flamegraphs below to see that that part of the code is taking up a significantly smaller portion of the overall time after the small cache was added.

Before flamegraph:
![before_direct_cache](https://user-images.githubusercontent.com/3280472/102251288-d9494500-3ec1-11eb-8365-c2191fcfcbf8.png)

After flamegraph:
![after_direct_cache](https://user-images.githubusercontent.com/3280472/102251306-df3f2600-3ec1-11eb-90ba-06566ffa81f7.png)