@brian-brazil 2020-12-15
BenchmarkPostingsForMatchers is the main query micro-benchmark.

> The cache that was recently added was only for the write path for the index, this is on the read path.

Could we do a more targeted cache there similar to what we did on that path? Allocating 24kB per touched block is quite a bit more than the usual per-query overheads. It'd also have more predictable performance.

Something also feels off with those allocation numbers, it shouldn't be taking us 90MB to do a select for a single series.