@csmarchbanks 2020-12-15
I put this into draft after taking a further look at the query path. The index reader is shared across all queries so the overhead is not per query, but that also means this is currently not safe as there is no locking upstream of this call. I will investigate further when I have some time.

I am a bit surprised nothing got detected in numerous tests with -race turned on.