@csmarchbanks 2021-05-20
Closing for now, I have some ideas around this but no plans to work on it in the near future.